use requestIdleCallback in react-import-remote
react-shortcuts: Remove prop-types as a dependency

PR #295 refactored react-shortcuts to use the modern context API which
means prop-types is no longer needed

Readme updates
