+++ packages/async/src/index.ts
@@ -4,3 +4,27 @@ export interface LoadProps<T> {
   id?(): string;
   load(): Promise<Import<T>>;
 }
+
+export enum DeferTiming {
+  Mount,
+  Idle,
+}
+
+export type RequestIdleCallbackHandle = any;
+
+export interface RequestIdleCallbackOptions {
+  timeout: number;
+}
+
+export interface RequestIdleCallbackDeadline {
+  readonly didTimeout: boolean;
+  timeRemaining: (() => number);
+}
+
+export interface WindowWithRequestIdleCallback {
+  requestIdleCallback: ((
+    callback: ((deadline: RequestIdleCallbackDeadline) => void),
+    opts?: RequestIdleCallbackOptions,
+  ) => RequestIdleCallbackHandle);
+  cancelIdleCallback: ((handle: RequestIdleCallbackHandle) => void);
+}
