+++ packages/react-async/src/index.ts
@@ -3,7 +3,7 @@ export {Prefetcher} from './Prefetcher';
 export {PrefetchRoute} from './PrefetchRoute';
 export {createAsyncComponent, AsyncComponentType} from './component';
 export {createAsyncContext, AsyncContextType} from './provider';
-export {DeferTiming} from './shared';
+export {DeferTiming} from '@shopify/async';
 
 export {AsyncAssetContext, AsyncAssetManager} from './context/assets';
 export {PrefetchContext, PrefetchManager} from './context/prefetch';
