+++ packages/react-async/src/shared.ts
@@ -3,8 +3,3 @@ export type Prefetchable<Props> =
       Prefetch: React.ComponentType<Props>;
     }
   | React.ComponentType<Props>;
-
-export enum DeferTiming {
-  Mount,
-  Idle,
-}
