+++ packages/react-graphql/src/index.ts
@@ -1,4 +1,4 @@
-export {DeferTiming} from '@shopify/react-async';
+export {DeferTiming} from '@shopify/async';
 export {Query} from './Query';
 export {Prefetch, Props as PrefetchProps} from './Prefetch';
 export {createAsyncQueryComponent} from './async';
