+++ packages/react-import-remote/src/ImportRemote.tsx
@@ -1,5 +1,10 @@
 import * as React from 'react';
 import {Preconnect} from '@shopify/react-html';
+import {
+  RequestIdleCallbackHandle,
+  WindowWithRequestIdleCallback,
+  DeferTiming,
+} from '@shopify/async';
 import load from './load';
 
 export interface Props<Imported = any> {
@@ -9,11 +14,31 @@ export interface Props<Imported = any> {
   onError(error: Error): void;
   getImport(window: Window): Imported;
   onImported(imported: Imported): void;
+  defer?: DeferTiming;
 }
 
 export default class ImportRemote extends React.PureComponent<Props, never> {
+  private idleCallbackHandle?: RequestIdleCallbackHandle;
+
+  componentWillUnmount() {
+    if (this.idleCallbackHandle != null && 'cancelIdleCallback' in window) {
+      (window as WindowWithRequestIdleCallback).cancelIdleCallback(
+        this.idleCallbackHandle,
+      );
+    }
+  }
+
   async componentDidMount() {
-    await this.loadRemote();
+    if (
+      this.props.defer === DeferTiming.Idle &&
+      'requestIdleCallback' in window
+    ) {
+      this.idleCallbackHandle = (window as WindowWithRequestIdleCallback).requestIdleCallback(
+        this.loadRemote,
+      );
+    } else {
+      await this.loadRemote();
+    }
   }
 
   async componentDidUpdate({source: oldSource}: Props) {
